+++ lib/carrierwave/base64/orm/mongoid.rb
@@ -1,23 +1,4 @@
-require 'mongoid'
+require "mongoid"
+require "carrierwave/base64/orm/base_db_adapter"
 
-module Carrierwave
-  module Base64
-    module Mongoid
-
-      def mount_base64_uploader(attribute, uploader_class, options = {})
-        mount_uploader attribute, uploader_class, options
-
-        define_method "#{attribute}=" do |data|
-          if data.present? && data.is_a?(String) && data.strip.start_with?("data")
-            super(Carrierwave::Base64::Base64StringIO.new(data.strip))
-          else
-            super(data)
-          end
-        end
-      end
-
-    end
-  end
-end
-
-Mongoid::Document::ClassMethods.send(:include, Carrierwave::Base64::Mongoid)
+Mongoid::Document::ClassMethods.send(:include, Carrierwave::Base64::BaseDBAdapter)
