+++ spec/orm/activerecord_spec.rb
@@ -1,8 +1,6 @@
 require "spec_helper"
 
-require "carrierwave/base64/orm/activerecord"
-
-RSpec.describe Carrierwave::Base64::ActiveRecord do
+RSpec.describe "carrierwave/base64/orm/activerecord" do
   describe ".mount_base64_uploader" do
     let(:uploader) { Class.new CarrierWave::Uploader::Base }
